[CVD3-555] fix UserController

+++ rest_service/src/main/java/ru/dexsys/rest_service/controller/UserController.java
@@ -2,11 +2,10 @@ package ru.dexsys.rest_service.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.dexsys.domain.entity.UserEntity;
 import ru.dexsys.domain.service.UserService;
 import ru.dexsys.rest_service.dto.UserDto;
 import ru.dexsys.rest_service.exception.NoSuchUserException;
@@ -26,7 +25,7 @@ public class UserController {
     private final UserDataDtoMapper userDataDtoMapper;
 
     @GetMapping("/users")
-    public HttpEntity<List<UserDto>> getUsers() {
+    public ResponseEntity<List<UserDto>> getUsers() {
         return ResponseEntity.ok(
                 userService.getUsers()
                         .stream()
@@ -36,47 +35,65 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public HttpEntity<UserDto> getUserById(@PathVariable Long id) {
-        return ResponseEntity.ok(
-                userService.getUser(id)
-                        .map(userDtoMapper)
-                        .orElseThrow(() -> NoSuchUserException.init(id))
-        );
+    public ResponseEntity<?> getUserById(@PathVariable Long id) {
+        if (userService.hasUser(id)) {
+            return ResponseEntity.ok(
+                    userService.getUser(id)
+                            .map(userDtoMapper)
+                            .orElseThrow(() -> NoSuchUserException.init(id))
+            );
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User #" + id + " not found");
+        }
     }
 
     @DeleteMapping("/users/{id}")
-    public HttpEntity<?> deleteUser(@PathVariable Long id) {
-        userService.removeUser(id);
-        return ResponseEntity.ok("User #" + id + " was removed from storage");
+    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+        if (userService.hasUser(id)) {
+            userService.removeUser(id);
+            return ResponseEntity.ok("User #" + id + " was removed from storage");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User #" + id + " not found");
+        }
     }
 
     @PostMapping(value = "/users", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public HttpEntity<UserDto> addUser(@RequestBody UserDto userDto) {
+    public ResponseEntity<UserDto> addUser(@RequestBody UserDto userDto) {
         userService.save(userDataDtoMapper.apply(userDto));
         return ResponseEntity.ok(userDto);
     }
 
     @PatchMapping(value = "/users/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public HttpEntity<UserDto> changeUser(
-            @RequestBody UserEntity userEntity,
-            @PathVariable Long id
+    public ResponseEntity<?> changeUser(
+            @RequestBody UserDto userDto
     ) {
-        userService.removeUser(id);
-        return ResponseEntity.ok(userDtoMapper.apply(userService.save(userEntity)));
+        if (userService.hasUser(userDto.getId())) {
+            userService.removeUser(userDto.getId());
+            userService.save(userDataDtoMapper.apply(userDto));
+            return ResponseEntity.ok(userDto);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User #" + userDto.getId() + " not found");
+        }
     }
 
     @PutMapping("/users/{id}")
-    public HttpEntity<UserDto> changeUserValues(
+    public ResponseEntity<UserDto> changeUserValues(
             @RequestHeader(required = false) Integer day,
             @RequestHeader(required = false) Integer month,
             @PathVariable Long id
     ) {
-        return ResponseEntity.ok(
-                userService.getUser(id)
-                        .map(user -> user.setDay(day))
-                        .map(user -> user.setMonth(month))
-                        .map(userDtoMapper)
-                        .orElseThrow(() -> NoSuchUserException.init(id))
-        );
+        if (day != null) {
+            userService.updateDay(id, day);
+        }
+        if (month != null) {
+            userService.updateMonth(id, month);
+        }
+        UserDto userDto = userService.getUser(id)
+                .map(userDtoMapper)
+                .orElseThrow(() -> NoSuchUserException.init(id));
+        return ResponseEntity.ok(userDto);
     }
 }
